rag/requirements.txt
fastapi
uvicorn
boto3
opensearch-py
python-dotenv
pydantic
pydantic-settings
requests
pytest
pytest-cov

